requirements.txt
twint
numpy
pandas
streamlit
matplotlib
textblob 
wordcloud
plotly
nltk
tensorflow-cpu
